Wrap into CommandLineTest
